updates from over the years... modified input file to support track numbering of output, more recent/accurate encoding, many other things i've forgotten

+++ ripdvd.pl
@@ -22,7 +22,32 @@ use strict;
 use diagnostics;
 use POSIX qw(tmpnam);
 
-my $trackbase = $ARGV[0] || "track";
+my $trackbase = "";
+my $wavonly = 0;
+my $device = "";
+if (($ARGV[0]) && (($ARGV[0] eq "--help") || ($ARGV[0] eq "-h"))) {
+   print qq(\n);
+   print qq(usage: ripdvd.pl [options] [trackbase]\n\n);
+   print qq(   options:\n);
+   print qq(      --wavonly, -w: rip audio to WAV format only. no mp3 encoding\n);
+   print qq(      --device, -d: use the specified device/iso.\n);
+   print qq(      --help, -h   : this help menu\n);
+   print qq(\n);
+   print qq(   trackbase:\n);
+   print qq(      base name for track if using default naming.\n);
+   print qq(      files rip as: trackbase-title##-chapter##\n);
+   print qq(\n);
+   exit;
+} elsif (($ARGV[0]) && (($ARGV[0] eq "--wavonly") || ($ARGV[0] eq "-w"))) {
+   $wavonly = 1;
+   $trackbase = $ARGV[1] || "track";
+} elsif (($ARGV[0]) && (($ARGV[0] eq "--device") || ($ARGV[0] eq "-d"))) {
+   $device = " -dvd-device '" . $ARGV[1] . "' ";
+} elsif ($ARGV[0]) {
+   $trackbase = $ARGV[0];
+} else {
+   $trackbase = "track";
+}
 
 my %dvd;
 $dvd{artist} = "";
@@ -30,8 +55,9 @@ $dvd{title} = "";
 $dvd{tno} = 0; # number of titles on dvd
 # $dvd{tno##} = 0; # chapter counts for each title
 $dvd{year} = "";
-@{$dvd{track}} = (); # array of title names
+@{$dvd{ref}} = (); # array of title names
 # @{$dvd{track##}} = (); # array of chapter names for each title
+# @{$dvd{label##}} = (); # array of chapter labels for each title
 #$dvd{torip}->{title,chapter} = (); # hash of "title,chapter" items to rip
 
 my $fixname;
@@ -75,12 +101,20 @@ my $disc_cycle;
 sub disc_cycle {
    my $user_input;
 
+   %dvd = ();
+   $dvd{artist} = "";
+   $dvd{title} = "";
+   $dvd{tno} = 0;
+   $dvd{aid} = 0;
+   $dvd{year} = "";
+   @{$dvd{ref}} = ();
+
    print "\n";
    print "---Insert DVD into drive and press enter to begin---\n";
    print "\n";
    $user_input = <STDIN>;  # FIXME: find cleaner way to get a key
    $user_input = 2;
-   print "getting CD info...\n";
+   print "getting DVD info...\n";
 
    probe_dvd();
 
@@ -121,46 +155,50 @@ sub probe_dvd {
 
    # first get the number of titles on the DVD
    $outfile = tmpnam();
-   $cmd = "mplayer dvd://1 -waveheader -vc null -vo null -aofile foo.wav -ao pcm -chapter 0-0 -af resample=44100,volume=0 > $outfile 2> /dev/null";
+   $cmd = "mplayer $device dvd:// -identify -frames 0 > $outfile 2> /dev/null";
    system($cmd);
 
    open STATUS, "$outfile" or die "cant open \"$outfile\": $!";
    while (<STATUS>) {
       $out = $_;
-      if ($_ =~ m/^There are (\d+) titles on this DVD/) {
+      if ($_ =~ m/^ID_DVD_TITLES=(\d+)$/) {
          $dvd{tno} = $1;
       }
+
+      if ($_ =~ m/^ID_DVD_TITLE_(\d+)_CHAPTERS=(\d+)$/) {
+         my $ref = "tno$1";
+         $dvd{$ref} = $2;
+
+         my $ref2 = "track$1";
+         for (my $j = 1; $j <= $dvd{$ref}; $j++) {
+            push @{$dvd{$ref2}}, "title $1 chapter $j";
+         }
+      }
+
+      if ($_ =~ m/^audio stream: \d+ format: [lpcmac3]+ \(stereo\) language: en aid: (\d+)\./) {
+#         print $out;
+         $dvd{aid} = $1;
+         last;
+      }
+
    }
    close STATUS;
    unlink $outfile or die "error unlinking \"$outfile\": $!";
 
-   # now find the number of chapters in each title
-   for (my $i = 1; $i <= $dvd{tno}; $i++) {
-      $outfile = tmpnam();
-      $cmd = "mplayer dvd://$i -waveheader -vc null -vo null -aofile foo.wav -ao pcm -chapter 0-0 -af resample=44100,volume=0 > $outfile 2> /dev/null";
-      system($cmd);
-
-      open STATUS, "$outfile" or die "cant open \"$outfile\": $!";
-      while (<STATUS>) {
-         $out = $_;
-         if ($_ =~ m/^There are (\d+) chapters in this DVD title/) {
-            my $ref = "tno$i";
-            $dvd{$ref} = $1;
-
-            my $ref2 = "track$i";
-            for (my $j = 1; $j <= $dvd{$ref}; $j++) {
-               push @{$dvd{$ref2}}, "title $i chapter $j";
-            }
-         }
-      }
-      close STATUS;
-      unlink $outfile or die "error unlinking \"$outfile\": $!";
-   }
 }
 
 my $print_dvd;
 sub print_dvd {
    my $dvd = shift;
+
+   if ($dvd->{aid})
+   {
+      print qq(stereo audio is in aid $dvd->{aid}\n);
+   }
+   else
+   {
+      print qq(no stereo audio found, using default\n);
+   }
 
    for (my $i = 1; $i <= $dvd->{tno}; $i++) {
       my $ref = "tno$i";
@@ -176,6 +214,7 @@ sub read_dvd_info {
    my $out;
    my $tnum;
    my $num;
+   my $tlabel;
    my $tmp;
 
    while (lc($user_input) ne "") {
@@ -197,14 +236,17 @@ sub read_dvd_info {
                $dvd->{title} = $1;
             } elsif ($out =~ m/^year (.*?)$/) {
                $dvd->{year} = $1;
-            } elsif ($out =~ m/^(\d+) (\d+) (.*?)$/) {
-               $tnum = $1; # title#
-               $num = $2;  # chapter#
-               $tmp = $3;  # chapter name
+            } elsif ($out =~ m/^(\d+) (\d+) ([^ ]+) (.*?)$/) {
+               $tnum = $1;    # title#
+               $num = $2;     # chapter#
+               $tlabel = $3;  # track label
+               $tmp = $4;     # chapter name
                my $ref = "$tnum,$num";
                my $ref2 = "track$tnum";
+               my $ref3 = "label$tnum";
 
                @{$dvd->{$ref2}}[$num-1] = $tmp;
+               @{$dvd->{$ref3}}[$num-1] = $tlabel;
                $dvd->{torip}->{$ref}++;
             } else {
                # ignore
@@ -229,13 +271,14 @@ sub edit_dvd_info {
    my $foo = "";
 
    %backup_dvd = %{$dvd};
-   @backup_tracks = @{$dvd->{track}};
+   @backup_tracks = @{$dvd->{ref}};
    while ((lc($user_input) ne "s") && (lc($user_input) ne "x")) {
       print "\n    [a] Artist  : $dvd->{artist}\n";
       print "    [t] Title   : $dvd->{title}\n";
       print "    [y] Year    : $dvd->{year}\n";
+      print "    [i] Audio   : $dvd->{aid}\n";
       print "\n    [c] Current Title: $curtitle\n";
-      print "\nrip ch# title\n";
+      print "\nrip ch# lbl title\n";
 
       my $num = 0;
       my $ref = "track$curtitle";
@@ -248,13 +291,18 @@ sub edit_dvd_info {
          } else {
             print "[n] ";
          }
-         print "[$num] $track\n";
+         print "[$num] [";
+         my $tlabel = @{$dvd->{"label$curtitle"}}[$num-1];
+         if ($tlabel && ($tlabel ne "")) {
+            print $tlabel;
+         }
+         print "] $track\n";
       }
       print "\n[g] toggle all\n";
       print "[f] read from file\n";
       print "[s] save\n";
       print "[x] abort and exit\n";
-      print "choice? ([s]/c/a/t/y/#/x/r#/g/f) ";
+      print "choice? ([s]/c/a/t/y/#/x/r#/l#/g/f) ";
 
       $user_input = <STDIN>;
       chomp($user_input);
@@ -287,6 +335,13 @@ sub edit_dvd_info {
          %{$dvd} = %backup_dvd;
          $dvd->{$ref} = \@backup_tracks;
       } elsif (($user_input =~ m/^r(\d+)$/) && ($1 <= $dvd->{$ref2})) {
+         $foo = "$curtitle,$1";
+         if ($dvd->{torip}->{$foo}) {
+            $dvd->{torip}->{$foo} = undef;
+         } else {
+            $dvd->{torip}->{$foo}++;
+         }
+      } elsif (($user_input =~ m/^l(\d+)$/) && ($1 <= $dvd->{$ref2})) {
          $foo = "$curtitle,$1";
          if ($dvd->{torip}->{$foo}) {
             $dvd->{torip}->{$foo} = undef;
@@ -338,7 +393,11 @@ sub rip_tag {
       my $arname = fixname($dvd->{artist});
       my $alname = fixname($dvd->{title});
       system("mkdir $arname");
-      system("mkdir $arname/$alname");
+      if ($wavonly) {
+         system("mkdir $arname/$alname-wav");
+      } else {
+         system("mkdir $arname/$alname");
+      }
 
       my $num;
       my $tnum;
@@ -361,30 +420,47 @@ sub rip_tag {
 
          my $t = "track$curtitle";
          my $track = @{$dvd->{$t}}[$i-1];
+         $t = "label$curtitle";
+         my $tlabel = @{$dvd->{$t}}[$i-1];
          my $cmd3 = $cmd2;
          if ((!$track) || ($track eq "")) {
             $cmd3 .= qq( -track="$num" -song="");
          } else {
             $cmd3 .= qq( -track="$num" -song="$track");
          }
-
-         my $fname = fixname(qq($dvd->{artist}-$dvd->{title}-$tnum-$num-$track.mp3));
+         if ((!$tlabel) || ($tlabel eq "")) {
+            $tlabel = "$tnum-$num";
+         }
+         my $fname = fixname(qq($dvd->{artist}-$dvd->{title}-$tlabel-$track.mp3));
 
          $cmd3 .= " $fname";
 
          $outfile = "$trackbase-title$tnum-chapter$num";
+
+         # rip wav audio from dvd
 #         $cmd = "mplayer -dvd $curtitle -vc dummy -vo null -hardframedrop -chapter $i-$i -ao pcm -really-quiet -nojoystick -nolirc -aid 128 -af resample=44100,volume=0 -waveheader -aofile '$fname.wav'";
-         $cmd = "mplayer -dvd $curtitle -vc dummy -vo null -hardframedrop -chapter $i-$i -ao pcm -nojoystick -nolirc -aid 128 -af volume=0 -waveheader -aofile '$fname.wav'";
+
+         my $aidstring = "";
+         if (defined $dvd->{aid})
+         {
+            $aidstring = "-aid $dvd->{aid}";
+         }
+
+         $cmd = "mplayer $device dvd://$curtitle -vc null -vo null -hardframedrop -chapter $i-$i -nojoystick -nolirc $aidstring -ao pcm:file='$fname.wav'";
          system($cmd);
 
-         # encode mp3
-         $cmd = "lame -h -b 192 '$fname.wav' '$fname'";
-         system($cmd);
+         if ($wavonly) {
+            system("mv '$fname.wav' $arname/$alname-wav");
+         } else {
+            # encode mp3
+            $cmd = "lame -h -b 320 '$fname.wav' '$fname'";
+            system($cmd);
 
-         # tag new mp3, move into subdir, delete .wav
-         system($cmd3);
-         system("mv '$fname' $arname/$alname");
-         system("rm -f '$fname.wav'");
+            # tag new mp3, move into subdir, delete .wav
+            system($cmd3);
+            system("mv '$fname' $arname/$alname");
+            system("rm -f '$fname.wav'");
+         }
       }
 
    }
